fix: guided tour once per session + corporate as root landing page
- Tour: add sessionStorage check so auto-open only triggers once per
  login session, not on every navigation to My Activity
- Landing: redirect / to /corporate (client request), preserve old
  landing page at /landing
- Update _redirects to avoid redirect loop
