+++ CHANGELOG.md
@@ -41,6 +41,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Toast scoping: informational toasts auto-dismiss after 4 seconds and all toasts clear on navigation (#151)
 - Fix `use-tournament-matches` and `submissions-table` importing toast directly from sonner, bypassing wrapper duration rules (#151)
 - Add inputMode="decimal" to distance input field for proper mobile keyboard (#133)
+- Guided tour only shows once per login session instead of on every navigation
+- Root URL `/` now serves the corporate landing page; old landing preserved at `/landing`
 
 ## [2.0.0] — 2026-04-01
