+++ public/_redirects
@@ -1,2 +1,2 @@
 /mfl-residential-wired.html /communities 301
-/mfl-landing-wired.html / 301
+/mfl-landing-wired.html /corporate 301
